Wrap up program counter chip
This ended up being pretty difficult. The key conceptual insight to get
it implemented was the realization that to implement mutually exclusive
conditionals I had to go backwards using multiplexors. For example:
if (a == 1)
  // do thing one
elsif (a == 3)
  // do thing two
else
  // do thing three

The way to do this was to multiplex the thing three result (i.e.
actually wire up the requirements to implement thing three) and the thing
two result, using the condition (a == 3) as the selector bit. If that
condition is met, the thing two result will be returned, otherwise it
will be the thing three result. Then, we just do the same with the thing
one result using the (a == 1) condition as the selector bit. As
mentioned before, a side effect of this approach is that the results
themselves actually have to be computed so we can wire them up to be
returned from the multiplexor.